requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.1
aniso8601==9.0.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorclass==2.2.2
crispy-bootstrap5==0.7
cryptography==41.0.6
defusedxml==0.7.1
Django==4.2.7
django-crispy-forms==2.0
django-extensions==3.2.3
django-filter==23.2
django-taggit==4.0.0
djangorestframework==3.14.0
dnspython==2.6.1
docopt==0.6.2
filelock==3.13.1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
gunicorn==21.2.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
Mako==1.3.0
Markdown==3.4.3
MarkupSafe==2.1.5
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
ngrok==1.0.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
pip-upgrader==1.4.15
pur==7.3.2
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyJWT==2.8.0
pymongo==4.6.3
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.1
python3-openid==3.2.0
python3-xlib==0.15
pytweening==1.2.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
sacremoses==0.1.1
safetensors==0.4.2
sentencepiece==0.1.99
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
SQLAlchemy==2.0.25
SQLAlchemy-serializer==1.4.1
sqlparse==0.4.4
sympy==1.12
terminaltables==3.1.10
tokenizers==0.15.1
torch==2.2.0
tqdm==4.66.2
transformers==4.37.2
triton==2.2.0
twilio==8.13.0
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
WTForms==3.1.2
yarl==1.9.4
zipp==3.15.0
